requirements.txt
python-telegram-bot==21.6
requests==2.28.1  # Solo si necesitas esta librería
